feat : 할인 정책에따른 최종 할인금액 받기

+++ src/main/java/domain/discount/CashPaymentMethodPolicy.java
@@ -12,7 +12,7 @@ public class CashPaymentMethodPolicy implements PaymentMethodDiscountPolicy {
     }
 
     @Override
-    public double discount(final int price) {
+    public double discount(final double price) {
         return price * DISCOUNT_RATE;
     }
 }
