+++ src/main/java/domain/discount/PaymentMethodDiscountPolicy.java
@@ -5,5 +5,5 @@ import domain.payment.PaymentMethod;
 public interface PaymentMethodDiscountPolicy {
     boolean support(PaymentMethod paymentMethod);
 
-    double discount(int price);
+    double discount(double price);
 }
